Move indexer notification to service layer
